Time: 48 ms (70.31%), Space: 7.8 MB (64.63%) - LeetHub

+++ 2443-sum-of-number-and-its-reverse/2443-sum-of-number-and-its-reverse.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-double reverse(int a){
+int reverse(int a){
     int reverse=0;
     while(a>0){
         reverse=reverse*10+(a%10);
